Move all effects back to components Effects can not be used outside body of functional component.
